Update the accuracy
